scraper/requirements.txt
# BoardGameGeek scraper dependencies
bgg-api>=1.1.16
python-dotenv>=1.0.0
requests>=2.31.0
